test(commands): isolate specialized variant regressions
