+++ tests/test_guard_command_extensions.py
@@ -174,28 +174,6 @@ def test_structured_core_rules_feed_runtime_classification(
 @pytest.mark.parametrize(
     "command",
     [
-        "git push origin +main",
-        "git push origin +refs/heads/main:refs/heads/main",
-    ],
-)
-def test_git_force_push_refspecs_feed_runtime_classification(command: str, tmp_path: Path) -> None:
-    payload = inspect_command(command, cwd=tmp_path, home_dir=tmp_path)
-    match = extract_sensitive_tool_action_request(
-        "Shell",
-        {"command": command},
-        cwd=tmp_path,
-        home_dir=tmp_path,
-    )
-
-    assert payload["status"] == "review"
-    assert payload["controlling_rule_id"] == "command.git.force-push"
-    assert match is not None
-    assert match.action_class == "git destructive command"
-
-
-@pytest.mark.parametrize(
-    "command",
-    [
         "git clean -nfdx",
         "git clean --dry-run -fdx",
         "git push origin main --force --dry-run",
